2.7.1 nova versão oficial
NEW) implementação de verificação de domínios em grace time no
registro.br.

UPDT) aprimoramento da estrutura BLOCK com objetivo de limpeza
automática.

FIX) correção da manipulação de dados na ferramenta ANALISE.

FIX) correção das constantes de tempo.

+++ src/net/spfbl/core/Server.java
@@ -394,12 +394,12 @@ public abstract class Server extends Thread {
     /**
      * Constante que representa a quantidade de tempo de um dia em milisegundos.
      */
-    public static final int HOUR_TIME = 1000 * 60 * 60;
+    public static final long HOUR_TIME = 1000L * 60L * 60L;
     
     /**
      * Constante que representa a quantidade de tempo de um dia em milisegundos.
      */
-    public static final int DAY_TIME = HOUR_TIME * 24;
+    public static final long DAY_TIME = HOUR_TIME * 24L;
         
     /**
      * Registra uma linha de LOG
