+++ src/net/spfbl/whois/AutonomousSystem.java
@@ -80,7 +80,7 @@ public class AutonomousSystem implements Serializable, Comparable<AutonomousSyst
      * @return verdadeiro se o registro atual expirou.
      */
     public boolean isRegistryExpired() {
-        int expiredTime = (int) (System.currentTimeMillis() - lastRefresh) / Server.DAY_TIME;
+        long expiredTime = (System.currentTimeMillis() - lastRefresh) / Server.DAY_TIME;
         return expiredTime > REFRESH_TIME;
     }
     
@@ -89,7 +89,7 @@ public class AutonomousSystem implements Serializable, Comparable<AutonomousSyst
      * @return verdadeiro se o registro atual está quase expirando.
      */
     public boolean isRegistryAlmostExpired() {
-        int expiredTime = (int) (System.currentTimeMillis() - lastRefresh) / Server.DAY_TIME;
+        long expiredTime = (System.currentTimeMillis() - lastRefresh) / Server.DAY_TIME;
         return expiredTime > (2 * REFRESH_TIME / 3); // Dois terços antes da expiração.
     }
